MAINT: Rename stream_python_file_by_line to stream_pyobject
