+++ setup.py
@@ -37,7 +37,7 @@ def configuration(parent_package='', top_path=None):
               'rows.c', 'tokenize.c.src',
               'conversions.c', 'str_to_int.c', 'str_to_double.c',
               'pow10table.c',
-              'stream_python_file_by_line.c',
+              'stream_pyobject.c',
               'field_types.c', 'dtoa_modified.c',]
     config.add_extension(
             'npreadtext._readtextmodule',
