+++ src/stream_pyobject.c
