+++ src/stream_pyobject.h
